Factor out meshchat.rs from main.rs
